move tests and update to rust 2018 idioms.
