--- identity_core/src/common/mod.rs
@@ -1,8 +1,0 @@
-#[macro_use]
-mod macros;
-
-mod object;
-mod timestamp;
-mod value;
-
-pub use self::{object::*, timestamp::*, value::*};
