+++ identity_core/src/common/timestamp.rs
@@ -76,48 +76,3 @@ impl TryFrom<&'_ str> for Timestamp {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    macro_rules! error_hack {
-        ($expr:expr) => {
-            match $expr {
-                Ok(_) => {}
-                Err(error) => panic!("{}", error),
-            }
-        };
-    }
-
-    #[test]
-    fn test_parse_valid() {
-        let original = "2020-01-01T00:00:00Z";
-        let timestamp = Timestamp::try_from(original).unwrap();
-
-        assert_eq!(timestamp.to_rfc3339(), original);
-
-        let original = "1980-01-01T12:34:56Z";
-        let timestamp = Timestamp::try_from(original).unwrap();
-
-        assert_eq!(timestamp.to_rfc3339(), original);
-    }
-
-    #[test]
-    #[should_panic = "premature end of input"]
-    fn test_parse_invalid_empty() {
-        error_hack!(Timestamp::try_from(""));
-    }
-
-    #[test]
-    #[should_panic = "input contains invalid characters"]
-    fn test_parse_invalid_bad_date() {
-        error_hack!(Timestamp::try_from("foo bar"));
-    }
-
-    #[test]
-    #[should_panic = "input contains invalid characters"]
-    fn test_parse_invalid_bad_fmt() {
-        error_hack!(Timestamp::try_from("2020/01/01 03:30:16"));
-    }
-}
